% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{match_adresses}
\alias{match_adresses}
\title{Match adresses data}
\format{Data frame with 1718 rows and 6 columns.
\describe{
  \item{league_gender}{Men or women's league.}
  \item{season_id}{ID for each season unique within league_gender. It consists of the last two digits of the season
  years, e.g., the season_id of season 2014/2015 equals 1415.}
  \item{match_id}{Official VBL match id. Unique within seasons.}
  \item{gym}{Official name of arena.}
  \item{adress}{ZIP code and city.}
  \item{lon}{Longitude of arena.}
  \item{lat}{Latitude of arena.}
}}
\source{
\url{https://www.volleyball-bundesliga.de/cms/home/1blf/1blf_clubsspieler/1blf_spielhallen.xhtml}
        \url{https://www.volleyball-bundesliga.de/cms/home/1blm/1blm_clubsspieler/1blm_spielhallen.xhtml}
}
\usage{
match_adresses
}
\description{
A dataset containing name, adress, and coordinates of the gym of each match in the German first division volleyball league between
2013/2014 and ending in 2018/2019.
}
\examples{
match_adresses
}
\keyword{datasets}
