% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{staff}
\alias{staff}
\title{Team staff data}
\format{A data frame with 1214 rows and 10 variables:
\describe{
  \item{league_gender}{Men or women's league.}
  \item{season_id}{ID for each season unique within league_gender. It consists of the last two digits of the season
  years, e.g., the season_id of season 2014/2015 equals 1415.}
  \item{team_id}{Team identifier which is unique across all teams and seasons.}
  \item{team_name}{Official name of the team. Note that VCO Berlin has a team in men's and women's league.}
  \item{firstname}{First name of the player}
  \item{lastname}{Last name of the player}
  \item{gender}{Player's gender: male or female.}
  \item{birthdate}{Date of birth}
  \item{position}{Person's role in the team, e.g., couch, statistician, or assistant couch.}
  \item{nationality}{Nationality of team member.}
}}
\source{
\url{https://www.volleyball-bundesliga.de/cms/home/1blf/1blf_archiv/1blf_tabellen_ergebnisse.xhtml}
        \url{https://www.volleyball-bundesliga.de/cms/home/1blm/1blm_archiv/1blm_tabellen_ergebnisse.xhtml}
}
\usage{
staff
}
\description{
A dataset containing attributes of all staff members of teams
in the German first division volleyball league for each season
starting in 2013/2014 and ending in 2018/2019.
}
\examples{
staff
}
\keyword{datasets}
