% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{team_adresses}
\alias{team_adresses}
\title{Team adresses data}
\format{Data frame with 137 rows and 8 columns.
\describe{
  \item{season_id}{ID for each season unique within league_gender. It consists of the last two digits of the season
  years, e.g., the season_id of season 2014/2015 equals 1415.}
  \item{league_gender}{Men or women's league.}
  \item{team_id}{Team identifier which is unique across all teams and seasons.}
  \item{team_name}{Official name of the team. Note that VCO Berlin has a team in men's and women's league.}
  \item{gym_adress}{ZIP code and city.}
  \item{max_spectator}{Maximum capacity}
  \item{lon}{Longitude of arena.}
  \item{lat}{Latitude of arena.}
}}
\source{
\url{https://www.volleyball-bundesliga.de/cms/home/1blf/1blf_clubsspieler/1blf_spielhallen.xhtml}
        \url{https://www.volleyball-bundesliga.de/cms/home/1blm/1blm_clubsspieler/1blm_spielhallen.xhtml}
}
\usage{
team_adresses
}
\description{
A dataset containing name, adress, and coordinates of the home gym of each team in the German first division volleyball league between
2013/2014 and ending in 2018/2019.
}
\examples{
team_adresses
}
\keyword{datasets}
